Restructure interface and remove implicit type conversions
The refactor to the interface primarily makes changes to require most
interfaces to use a Triangle instead of needing to pass potentially up
to 6 vectors. Also functions will no longer return functions as this is
a bit too complex with little to no benefit.

Furthermore, with the previously added compiler flags, dangerous type conversion is
being reduced especially in the draw_face function such that dangerous
conversions are minimized.